// src/generated/model/bostotteFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {JsonOkonomiOpplysningUtbetaling} from "./jsonOkonomiOpplysningUtbetaling";
import type {JsonBostotteSak} from "./jsonBostotteSak";

export interface BostotteFrontend {
    bekreftelse?: boolean;
    samtykke?: boolean;
    utbetalinger?: JsonOkonomiOpplysningUtbetaling[];
    saker?: JsonBostotteSak[];
    stotteFraHusbankenFeilet?: boolean;
    samtykkeTidspunkt?: string;
}